+++ broker/src/main/java/de/fraunhofer/ids/framework/broker/BrokerService.java
@@ -5,7 +5,12 @@ import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
 
-import de.fraunhofer.iais.eis.*;
+import de.fraunhofer.ids.framework.messaging.protocol.multipart.MultipartRequestBuilder;
+import de.fraunhofer.iais.eis.DynamicAttributeToken;
+import de.fraunhofer.iais.eis.QueryLanguage;
+import de.fraunhofer.iais.eis.QueryScope;
+import de.fraunhofer.iais.eis.QueryTarget;
+import de.fraunhofer.iais.eis.Resource;
 import de.fraunhofer.iais.eis.ids.jsonld.Serializer;
 import de.fraunhofer.ids.framework.config.ClientProvider;
 import de.fraunhofer.ids.framework.config.ConfigContainer;
@@ -14,14 +19,19 @@ import de.fraunhofer.ids.framework.messaging.protocol.http.IdsHttpService;
 import de.fraunhofer.ids.framework.messaging.protocol.multipart.MultipartRequestBuilder;
 import de.fraunhofer.ids.framework.messaging.util.IdsMessageUtils;
 import lombok.extern.slf4j.Slf4j;
+import okhttp3.Call;
+import okhttp3.Callback;
 import okhttp3.MultipartBody;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
-import okio.Buffer;
-import org.apache.commons.fileupload.FileUploadException;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
+
+import java.io.IOException;
+import java.net.URI;
+import java.util.ArrayList;
+import java.util.List;
 
 import static de.fraunhofer.ids.framework.messaging.util.RequestUtils.logRequest;
 
@@ -50,8 +60,10 @@ public class BrokerService implements IDSBrokerService {
      * @param tokenProvider  providing DAT Token for RequestMessage
      * @param idsHttpService providing sending capabilities
      */
-    public BrokerService( ConfigContainer container, ClientProvider provider, DapsTokenProvider tokenProvider,
-                          IdsHttpService idsHttpService ) {
+    public BrokerService( final ConfigContainer container,
+                          final ClientProvider provider,
+                          final DapsTokenProvider tokenProvider,
+                          final IdsHttpService idsHttpService) {
         this.container = container;
         this.clientProvider = provider;
         this.tokenProvider = tokenProvider;
@@ -63,7 +75,7 @@ public class BrokerService implements IDSBrokerService {
      * {@inheritDoc}
      */
     @Override
-    public Response removeResourceFromBroker( String brokerURI, Resource resource )
+    public Response removeResourceFromBroker( final String brokerURI, final Resource resource )
             throws IOException, DapsTokenManagerException {
         logBuildingHeader();
 
@@ -83,7 +95,7 @@ public class BrokerService implements IDSBrokerService {
      * {@inheritDoc}
      */
     @Override
-    public Response updateResourceAtBroker( String brokerURI, Resource resource ) throws
+    public Response updateResourceAtBroker( final String brokerURI, final Resource resource ) throws
             IOException,
             DapsTokenManagerException {
         logBuildingHeader();
@@ -102,12 +114,9 @@ public class BrokerService implements IDSBrokerService {
 
     /**
      * {@inheritDoc}
-     *
-     * @return
      */
     @Override
-    public Response unregisterAtBroker( String brokerURI )
-            throws IOException, DapsTokenManagerException, ClaimsException, FileUploadException {
+    public Response unregisterAtBroker( final String brokerURI ) throws IOException, DapsTokenManagerException {
         logBuildingHeader();
 
         var securityToken = getDat();
@@ -123,10 +132,9 @@ public class BrokerService implements IDSBrokerService {
 
     /**
      * {@inheritDoc}
-     * @return
      */
     @Override
-    public Response updateSelfDescriptionAtBroker( String brokerURI ) throws
+    public Response updateSelfDescriptionAtBroker( final String brokerURI ) throws
             IOException, DapsTokenManagerException {
         logBuildingHeader();
 
@@ -144,7 +152,6 @@ public class BrokerService implements IDSBrokerService {
 
     /**
      * {@inheritDoc}
-     * @return
      */
     @Override
     public List<Response> updateSelfDescriptionAtBrokers( List<String> brokerURIs ) throws
@@ -171,8 +178,11 @@ public class BrokerService implements IDSBrokerService {
      * {@inheritDoc}
      */
     @Override
-    public Response queryBroker( String brokerURI, String query, QueryLanguage queryLanguage, QueryScope queryScope,
-                                 QueryTarget queryTarget ) throws IOException, DapsTokenManagerException {
+    public Response queryBroker( final String brokerURI,
+                                 final String query,
+                                 final QueryLanguage queryLanguage,
+                                 final QueryScope queryScope,
+                                 final QueryTarget queryTarget ) throws IOException, DapsTokenManagerException {
         logBuildingHeader();
 
         var securityToken = getDat();
@@ -197,18 +207,10 @@ public class BrokerService implements IDSBrokerService {
      *
      * @throws IOException if requestBody cannot be sent
      */
-    @Deprecated( forRemoval = true )
-    private Response sendBrokerMessage( String brokerURI, RequestBody requestBody ) throws IOException {
-        Request request = new Request.Builder().url(brokerURI).post(requestBody).build();
-        try {
-            final Request copy = request.newBuilder().build();
-            final Buffer buffer = new Buffer();
-            copy.body().writeTo(buffer);
-            System.out.println(buffer.readUtf8());
-        } catch( final IOException e ) {
-        }
-
-        var response = clientProvider.getClient().newCall(request).execute();
+    private Response sendBrokerMessage( final String brokerURI, final RequestBody requestBody ) throws IOException {
+        var response = clientProvider.getClient().newCall(
+                new Request.Builder().url(brokerURI).post(requestBody).build()
+        ).execute();
 
         if( !response.isSuccessful() ) {
             log.warn("Response of the Broker wasn't successful!");
@@ -217,32 +219,69 @@ public class BrokerService implements IDSBrokerService {
         return response;
     }
 
+    /**
+     * Get the ID of the connector
+     *
+     * @return Connector-ID URI
+     */
     @NotNull
     private URI getConnectorId() {
         return container.getConnector().getId();
     }
 
+    /**
+     * Get a new DAT from the DAPS
+     *
+     * @return DAT, returned by the DAPS for the Connector
+     *
+     * @throws ConnectorMissingCertExtensionException Something went wrong with the Certificate of the Connector
+     * @throws DapsConnectionException                The DAPS is not reachable (wrong URL, network problems..)
+     * @throws DapsEmptyResponseException             The DAPS didn't return the expected response (maybe DAPS internal Problem?)
+     */
     private DynamicAttributeToken getDat()
             throws ConnectorMissingCertExtensionException, DapsConnectionException, DapsEmptyResponseException {
         return tokenProvider.getDAT();
     }
 
-    @Deprecated( forRemoval = true )
-    private MultipartBody getMessageBody( String header, String payload ) {
+    /**
+     * Build the MultiPart message with header and payload
+     *
+     * @param header  Header of the MultiPart message
+     * @param payload Payload of the MultiPart message
+     *
+     * @return Generated MultiPart Message
+     */
+    private MultipartBody getMessageBody( final String header, final String payload ) {
         return MessageBuilder.buildRequestBody(header, payload);
     }
 
+    /**
+     * Log info about starting to build the header
+     */
     private void logBuildingHeader() {
         log.debug("Building message header");
     }
 
-    @Deprecated( forRemoval = true )
-    private void logSendingMessage( String brokerURI ) {
+    /**
+     * Log info about starting to send the message to the Broker
+     *
+     * @param brokerURI URI of the Broker
+     */
+    private void logSendingMessage( final String brokerURI ) {
         log.debug(String.format("Sending message to %s", brokerURI));
     }
 
+    /**
+     * Serializes a given Resource
+     *
+     * @param resource The Resource to be serialized
+     *
+     * @return String, the serialized Resource
+     *
+     * @throws IOException Something went wrong reading the given Resource
+     */
     @NotNull
-    private String serializeResource( Resource resource ) throws IOException {
+    private String serializeResource( final Resource resource ) throws IOException {
         return SERIALIZER.serialize(resource);
     }
 }