+++ broker/src/main/java/de/fraunhofer/ids/framework/broker/IDSBrokerService.java
@@ -1,8 +1,5 @@
 package de.fraunhofer.ids.framework.broker;
 
-import java.io.IOException;
-import java.util.List;
-
 import de.fraunhofer.iais.eis.QueryLanguage;
 import de.fraunhofer.iais.eis.QueryScope;
 import de.fraunhofer.iais.eis.QueryTarget;
@@ -11,6 +8,9 @@ import de.fraunhofer.ids.framework.daps.ClaimsException;
 import de.fraunhofer.ids.framework.daps.DapsTokenManagerException;
 import okhttp3.Response;
 import org.apache.commons.fileupload.FileUploadException;
+
+import java.io.IOException;
+import java.util.List;
 
 /**
  * Interface for Communication with IDS Brokers, implemented by {@link BrokerService}