+++ messaging/src/main/java/de/fraunhofer/ids/framework/messaging/endpoint/EndpointService.java
@@ -1,7 +1,5 @@
 package de.fraunhofer.ids.framework.messaging.endpoint;
 
-import javax.servlet.http.HttpServletRequest;
-
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +8,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.mvc.method.RequestMappingInfo;
 import org.springframework.web.servlet.mvc.method.annotation.RequestMappingHandlerMapping;
+
+import javax.servlet.http.HttpServletRequest;
 
 /**
  * Service for modifying the route of {@link MessageController} in a running application